更新:資料庫

+++ GraduationProject_MVC/GraduationProject/DTOs/OrderUpdateDTO.cs
@@ -23,7 +23,7 @@ namespace GraduationProject.DTOs
 
         public string? DeliveryAddress { get; set; }
 
-        public int? ShippingCost { get; set; }
+        public decimal? ShippingCost { get; set; }
 
         public DateTime? DeliveryTime { get; set; }
